Create different components for SignUp and Login forms
Creating two different components is necessary because of the different
number of inputs we have for the SignUp and the Login. Also a useState
was added to AuthPage.js in order to render both the new Login and
Signup Form. Lastly the url for requests was changed and now our own API
 is used.